Made a start on defining shapes and stuff for use with the prototypes.
Shapes are defined by the same name in JSON files stored in settings, please note that the minimal-json library is included and if you're getting errors whilst trying to run it you'll have to include that jar.

Also note that only SquareBumper.json has the correct definition, all the other json files are copies of that and will need to be modified as such

Basically shapes have settings defined in two primary components: PhysicsDef and VisualDef.

PhysicsDef is a wrapper for easy modification of placement of the MIT physics objects, in the SquareBumper.json you can see that the format is as such [type, initialcoord, other], where initialcoord is where the shape should start from and other defined radius or destination for the shape.

The numbers range from 0-100 as they're percent based over the location they occupy. For example if this bumper is at x=10, y=10  then 0, 0 in the PhysicsDef means the top left corner of that grid, 100, 100 means the bottom right.

VisualDef is a wrapper to provide a way to actually draw the object without having to manually define exactly what each shape is, it's not implemented yet but on the view it'll be able to read Square, Circle and Triangle types and display them according to the stored coordinates

VisualShape is just an object representation of each entry in VisualDef, it stores the shape type, colour and an array of ints which will be interpreted as whatever the shape type defines.

Using this way we can seperate physical definitions from visual definitions and reduce long winded manual definitions in code. By doing it this way you can easily define shapes and also reload shape characteristics on the fly if desired. It also vastly reduces clutter in code.