Added some intellij files it was complaining about.
Decided to make the ball with the AbstractGizmo even though we agreed that it'd make more sense to have it different, i initially implemented it as a new object but then i realised i couldn't use the .json reader model with it and i'd need to hardcode the MIT circle and visual objects which seemed a bit redundant, repeated code etc.

It seems to work perfectly and by changing the inherited constructors in the Gizmos to use either ints or doubles we can actually restrict value placement to the grid squares but still allow the BallActor to go free.

For whoever is doing the ball movement, have a look at the timer in TestView, it kinda illustrates how the ball would move, note we'd probs need to add a call from the timer -> wherever we're going to be doing physics updates to put the physics code in.

Also changed the absorber to be allowed to have an arbritrary height, before it was locked to 1L