NOTE YOU SHOULD PROBABLY READ THIS
changed ALL parts of the model to use L values, ie a double ranging from (0 to 19) because using pixels to calculate the gravity/friction whatever was actually agonising and needless effort.

Instead everything on the model uses L values including velocity which is nice because it's easy to understand now, no more weird multiplications.

The ONLY PLACE pixels are used are when its' drawn it uses AffineTransform to scale up all shapes to be visible.

Actual functionality isn't affected, just makes it a lot easier to debug and figure out why the ball is going at 4000 Ls per second