Added better labelled settings such as using the standard FPS notation for frame rate and using the reciprocal of that to get the milliseconds for the Timer.
Note will likely need to figure out a better way for rotating/moving dynamic gizmos so they're not bound to the drawcode. Right now a gizmo is updated when its' shape is requested which does actually happen on every tick iteration but i dunno, seems a bit odd.

Flippers still being the worst