Tried to tidy collision code up a bit more, it was inconsistent between high-fps and low-fps.
previously if you used a low fps setting it would bug out and start glitching into other shapes but this was an issue with applying the gravity improperly or something, it seems to be fixed now.

Also i changed it back to the old way of gizmos calling their listeners without using weird static methods in the project manager.

Added the gizmo Connect command to the existing FileManager, seems to work fine.