Made a start on the test cases, had to modify the flipper a bit due to the x position being a bit weird with the square checking.
Added a square checking method in ProjectManager to allow checking to see if a gizmo can be placed into a square, ie if a gizmo is already occupying that square. it could probably be simplified a bit but that was the easiest way i could figure out how to do it.

Slightly modified FileManager to allow Absorbers of variable width and height